[mattwynne+jbpros] Start to decouple web app from core domain
We're retro-fitting a scenario that needs us to be able to inject pages
to be rendered. This has caused us to have to break the dependency
between the sinatra app and the pages being loaded from disk.

Eventually we hope to be able to use the web app or core domain easily
from tests without these dependencies getting in the way. This is a
start.